requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.6
fastapi==0.100.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
lxml==4.9.3
Pillow==10.0.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pydantic==2.1.1
pydantic_core==2.4.0
pypdf @ git+https://github.com/py-pdf/PyPDF2.git@c239073d64b3c21b1737f4c15267038c6387710d
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
reportlab==3.6.12
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1